requirements.txt
boto3
botocore
python-dotenv
pexpect